+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-beta.1] - 2025-05-22
+
+## Changed
+
+- Updated the frontend to work with the new backend
+- Core features fixed to work quite stable: Official first beta version
+
 ## [1.1.2-alpha.4] - 2025-05-20
 
 ## Added